refactor messenger drawer
